lint

+++ count/count_objects.cc
@@ -1,24 +1,22 @@
 #include "count/count_objects.h"
+
+#include <vector>
 #include "common.h"
 
-long long int count_objects_size(std::string bucket_name, gcs::Client* client) {
-
-    long long int count = 0;
+int64_t count_objects_size(std::string bucket_name, gcs::Client* client) {
+    int64_t count = 0;
     for (auto& object_metadata : client ->ListObjects(bucket_name)) {
         if (!object_metadata) {
-            std::cerr << "Error in getting object" << std::endl;              
+            std::cerr << "Error in getting object" << std::endl;
             exit(1);
         }
-
         count += object_metadata->size();
     }
-
     return count;
 }
 
-long long int count_objects_in_bucket(std::string bucket_name, gcs::Client* client) {
-
-    long long int count = 0;
+int64_t count_objects_in_bucket(std::string bucket_name, gcs::Client* client) {
+    int64_t count = 0;
     for (auto& object_metadata : client ->ListObjects(bucket_name)) {
         if (!object_metadata) {
             std::cerr << "Error in getting object" << std::endl;
@@ -31,17 +29,17 @@ long long int count_objects_in_bucket(std::string bucket_name, gcs::Client* clie
     return count;
 }
 
-long long int count_objects(gcs::Client* client, bool size) {
+int64_t count_objects(gcs::Client* client, bool size) {
     std::vector<std::string> bucket_prefixes = {
-        "frontend", "adservice", "cartservice", "checkoutservice", "currencyservice", 
-        "emailservice", "paymentservice", "productcatalogservice", "recommendationservice", 
+        "frontend", "adservice", "cartservice", "checkoutservice", "currencyservice",
+        "emailservice", "paymentservice", "productcatalogservice", "recommendationservice",
         "rediscart", "shippingservice",
         "dyntraces", "tracehashes"
     };
 
-    long long int count = 0;
+    int64_t count = 0;
 
-    std::vector<std::future<long long int>> response_futures;
+    std::vector<std::future<int64_t>> response_futures;
 
     for (auto ele : bucket_prefixes) {
         auto bucket_name = ele + BUCKETS_SUFFIX;
@@ -52,7 +50,7 @@ long long int count_objects(gcs::Client* client, bool size) {
         }
     }
 
-    for (int i = 0; i < response_futures.size(); i++) {
+    for (int64_t i = 0; i < response_futures.size(); i++) {
         auto res = response_futures[i].get();
         count += res;
     }
